Add difference reproduction tooling (#2)
* Add O0,O3 unmutated for comparison, refactor

* Refactor rmdir in MutateCSmithProvider

* Add comparison with unmutated for all opt levels

Instead of comparing with unmutated versions of just O0 and O3, this
commit automatically compares with all given opt levels. In addition,
file names for each seed directory are more intuitive, and most
human-unreadable files are made to be removed when not in debug mode.

* Remove --debug for --repro option

* Add batch repro script

* Add fn_info.json to seed artifact

This commit generates fn_info.json for each kept program seed directory.
Now kept seeds also retain the chosen function's return type and
argument types so that future repro scripts don't have to reparse the C
or IR files. Minor edits were also made to improve code clarity and
artifact naming.

* Add iterative verification scripts

* Add early exits for missing args to finditer.sh

+++ sem/fuzzing.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python3
 import contextlib
+from glob import glob
+import json
 import logging
 import os
 import re
@@ -53,6 +55,16 @@ def rmdir(dir: str):
         shutil.rmtree(dir, True)
 
 
+def save_seed(p: Program, program_seed: int):
+    """Rename a program's data dir it to program seed and return the path."""
+
+    data_dir = p.data_dir
+    dest = os.path.join(data_dir.rsplit("/", 1)[0], str(program_seed))
+    shutil.rmtree(dest, ignore_errors=True)
+    os.rename(data_dir, dest)
+    return dest
+
+
 @dataclass
 class Experiment:
     """Class for testing compiler optimization passes."""
@@ -67,6 +79,8 @@ class Experiment:
     randomizer: Randomizer
     timeout: int
     debug: bool
+    keep_data: bool
+    dump_regs: bool
 
     def __post_init__(self):
         global is_debug
@@ -92,6 +106,8 @@ class Experiment:
             self._programs = self.program_provider.get(self)
         except Exception as e:
             log.error("Program generation exception", exc_info=True)
+            if is_debug:
+                print(e)
             return (RunStatus.RUN_GEN_EXC, program_seed)
 
         assert len(self._programs) >= 2
@@ -107,7 +123,7 @@ class Experiment:
                 emulator = emu_info[0]
                 self.randomizer.update(emulator, self.context)
 
-            if self.debug:
+            if self.dump_regs:
                 print("Initial register dump:")
                 self._dump_registers()
 
@@ -135,17 +151,19 @@ class Experiment:
                     return (RunStatus.RUN_EMU_EXC, program_seed)
 
             self._diff = self._diff_vars()
-            if self.debug:
+            if self.dump_regs:
                 print("Final register dump:")
                 self._dump_registers()
             if len(self._diff):
-                data_dir = self._programs[0].data_dir
-                dest = os.path.join(data_dir.rsplit("/", 1)[0], str(program_seed))
-                rmdir(dest)
-                os.rename(data_dir, dest)
+                dest = save_seed(self._programs[0], program_seed)
+                with open(os.path.join(dest, "diff.txt"), "w") as f:
+                    f.write(str(self.make_diff_table()))
                 return (RunStatus.RUN_DIFF, program_seed)
 
-        rmdir(self._programs[0].data_dir)
+        if self.keep_data:
+            save_seed(self._programs[0], program_seed)
+        else:
+            rmdir(self._programs[0].data_dir)
         return (RunStatus.RUN_OK, program_seed)
 
     def _diff_vars(self, vars: list[Variable] = []) -> dict[Variable, list[bytes]]:
@@ -155,6 +173,10 @@ class Experiment:
         diff = {}
         for var in vars:
             values = [var.get(emu) for emu in self._emulators]
+            if self.program_provider.name == "mutate-csmith":
+                if any(values[0] != value for value in values[1 : len(values) // 2]):
+                    # Ignore cases where CSmith produced the difference
+                    continue
             if all(values[0] == value for value in values[1:]):
                 continue
             diff[var] = values
@@ -162,7 +184,12 @@ class Experiment:
 
     def _dump_registers(self):
         vars = self.context.variables
-        table = {var: [var.get(emu) for emu in self._emulators] for var in vars}
+        table = {
+            var: [var.get(emu) for emu in self._emulators]
+            for var in vars
+            # TODO: ignore SIMD registers ISA-agnostically
+            if "ymm" not in var.name
+        }
         print(self.make_diff_table(table))
 
     def make_diff_table(self, diff: dict[Variable, list[bytes]] = {}):
@@ -349,7 +376,7 @@ class IRFuzzerProvider(ProgramProvider):
             )
             ir_ll_path = os.path.join(tmpdir, "out.ll")
             subprocess.run(["llvm-dis", ir_bc_path, "-o", ir_ll_path])
-            name, ret_ty, arg_tys = self.choose_ir_fn(experiment, ir_ll_path)
+            name, ret_ty, arg_tys = self._choose_ir_fn(experiment, ir_ll_path)
 
             programs: list[Program] = []
             for opt_level in experiment.opt_levels:
@@ -397,7 +424,7 @@ class IRFuzzerProvider(ProgramProvider):
             # TODO: llc: check if memcpy and other builtins are used
             return programs
 
-    def choose_ir_fn(
+    def _choose_ir_fn(
         self,
         experiment: Experiment,
         ir_ll_path: str,
@@ -418,7 +445,8 @@ class IRFuzzerProvider(ProgramProvider):
             fn_name = m.group("fn_name")
             ret_ty = self._parse_arg_tys(m.group("ret_ty"))[0]
             arg_list = m.group("arg_list")
-            if fn_name in ["memcpy", "memset"] or fn_name.startswith("safe_"):
+            # Choose CSmith generated function only
+            if not fn_name.startswith("func_"):
                 continue
             last_generated_fn = (fn_name, ret_ty, self._parse_arg_tys(arg_list))
             if experiment.randomizer.choice([True, False]):
@@ -427,7 +455,7 @@ class IRFuzzerProvider(ProgramProvider):
                 return last_generated_fn
         if not last_generated_fn:
             raise RuntimeError("No generated functions found")
-        # Settle for empty parameter list
+        # Settle for potentially empty parameter list
         return last_generated_fn
 
     def _parse_arg_tys(self, arg_list: str) -> list[str]:
@@ -494,133 +522,215 @@ class IRFuzzerProvider(ProgramProvider):
 
 
 class MutateCSmithProvider(CSmithProvider, IRFuzzerProvider):
-    def get(self, experiment: Experiment) -> list[Program]:
-        with contextlib.nullcontext(
-            tempfile.mkdtemp(prefix=experiment.output_dir)
-        ) as tmpdir:
-            # TODO: organize some of the steps into inherited functions to remove redundant code
-            # Generate source
-            source_c_path = os.path.join(tmpdir, "out.c")
-            source_bc_path = os.path.join(tmpdir, "out.bc")
-            source_ll_path = os.path.join(tmpdir, "out.ll")
-            self.gen_csmith_program(experiment, source_c_path)
+    """Provide programs generated by CSmith and mutated by IRFuzzer.
 
-            subprocess.run(
+    Allows override certain parameters via environment variables
+    - NUM_MUTATE: controls number of mutations
+    - FN_INFO: override which function gets chosen and its signature
+    """
+
+    # Allow overriding number of iterations; useful for finding the
+    # difference-causing iteration
+    NUM_MUTATE = int(os.environ.get("NUM_MUTATE", "100"))
+
+    def get(self, experiment: Experiment) -> list[Program]:
+        """Get a list of programs (O*, O* unmutated). Deletes tmpdir if error occurs."""
+        tmpdir = tempfile.mkdtemp(prefix=experiment.output_dir)
+        try:
+            return self._get(experiment, tmpdir)
+        except:
+            rmdir(tmpdir)
+            raise
+
+    def _get(self, experiment: Experiment, cwd: str) -> list[Program]:
+        """Get a list of programs (O*, O* unmutated)."""
+
+        # Generate CSmith sources
+        csmith_c_path = os.path.join(cwd, "csmith.c")
+        csmith_ll_path = os.path.join(cwd, "csmith.ll")
+        self.gen_csmith_program(experiment, csmith_c_path)
+        self._clang(csmith_c_path, csmith_ll_path)
+
+        # Mutate .ll
+        # NOTE: This has to be out.bc since that's also what IRFuzzer outputs into.
+        irf_bc_path = os.path.join(cwd, "out.bc")
+        subprocess.run(["llvm-as", csmith_ll_path, "-o", irf_bc_path], check=True)
+        self._mutate(irf_bc_path, cwd, experiment)
+
+        # Disassemble to choose a function
+        irf_ll_path = os.path.join(cwd, "out.ll")
+        subprocess.run(["llvm-dis", irf_bc_path, "-o", irf_ll_path], check=True)
+
+        # Get override from env; useful for reproduction independent of NUM_MUTATE.
+        fn_info = os.environ.get("FN_INFO", "")
+        if fn_info:
+            fn_info = tuple(json.loads(fn_info))
+        else:
+            fn_info = self._choose_ir_fn(experiment, csmith_ll_path)
+        with open(os.path.join(cwd, "fn_info.json"), "w") as f:
+            f.write(json.dumps(fn_info))
+
+        # Make unmutated programs for comparison
+        programs: list[Program] = []
+        arch = experiment.context.arch
+        triple = f"{arch if arch != 'x86' else 'x86_64'}--"
+        for opt_level in experiment.opt_levels:
+            csmith_elf_path = os.path.join(cwd, f"csmith.{opt_level}.elf")
+            self._compile_elf(opt_level, triple, csmith_ll_path, csmith_elf_path)
+            programs.append(
+                self._make_program(
+                    csmith_elf_path, f"O{opt_level}", fn_info, cwd, False
+                )
+            )
+
+        # Compile .o with EMI global
+        emi_ll_path = os.path.join(cwd, "emi_false.ll")
+        emi_o_path = os.path.join(cwd, "emi_false.o")
+        with open(emi_ll_path, "w") as f:
+            f.write("@emi_false = global i1 0")
+        subprocess.run(
+            [
+                "llc",
+                "-filetype=obj",
+                f"-mtriple={triple}",
+                emi_ll_path,
+                f"-o",
+                emi_o_path,
+            ],
+            check=True,
+        )
+
+        for opt_level in experiment.opt_levels:
+            elf_path = os.path.join(cwd, f"{opt_level}.elf")
+
+            self._compile_elf(opt_level, triple, irf_bc_path, elf_path, [emi_o_path])
+            programs.append(
+                self._make_program(
+                    elf_path,
+                    f"O{opt_level}*",
+                    fn_info,
+                    cwd,
+                    True,
+                )
+            )
+
+        def rm_glob(pattern):
+            files = glob(os.path.join(cwd, pattern))
+            for file in files:
+                os.remove(file)
+
+        if not is_debug:
+            os.remove(emi_ll_path)
+            rm_glob("*.o")
+            rm_glob("*.elf")
+            rm_glob("*.bc")
+
+        return programs
+
+    def _clang(self, source: str, output: str, opt_level: str = "0"):
+        """Use clang to compile source into ELF or IR file."""
+        clang_args = [
+            "clang",
+            f"-O{opt_level}",  # don't test middle-end by default
+            f"-I{self.CSMITH_RUNTIME}",
+            "-nostdlib",
+            "-ffreestanding",
+            "-fno-builtin",
+            source,
+            "-o",
+            output,
+        ]
+
+        if ".ll" in output:
+            clang_args.extend(
                 [
-                    "clang",
                     "-S",
                     "-emit-llvm",
-                    "-O0",
                     "-Xclang",
                     "-disable-O0-optnone",
-                    f"-I{self.CSMITH_RUNTIME}",
-                    "-nostdlib",
-                    "-ffreestanding",
-                    "-fno-builtin",
-                    source_c_path,
-                    "-o",
-                    source_ll_path,
-                ],
-                stderr=subprocess.DEVNULL,
-            )
-            subprocess.run(["llvm-as", source_ll_path])
-
-            irfuzzer_env = os.environ.copy()
-            irfuzzer_env["NUM_MUTATE"] = str(self.MUTATION_ITERS)
-            mutation = subprocess.run(
-                ["MutatorDriver", source_bc_path, str(experiment.randomizer.get())],
-                env=irfuzzer_env,
-                cwd=tmpdir,
-                stderr=subprocess.DEVNULL,
-            )
-            if mutation.returncode != 0:
-                rmdir(tmpdir)
-                raise RuntimeError(mutation.stderr)
-
-            subprocess.run(["llvm-dis", source_bc_path])
-            try:
-                fn_name, ret_ty, arg_tys = self.choose_ir_fn(experiment, source_ll_path)
-                with open(os.path.join(tmpdir, "chosen_function.txt"), "w") as f:
-                    f.write(fn_name)
-            except:
-                rmdir(tmpdir)
-                raise
-            programs: list[Program] = []
-
-            emi_ll_path = os.path.join(tmpdir, "emi_false.ll")
-            emi_o_path = os.path.join(tmpdir, "emi_false.o")
-            with open(emi_ll_path, "w") as f:
-                f.write("@emi_false = global i1 0")
-            arch = experiment.context.arch
-            llc_triple = f"-mtriple={arch if arch != 'x86' else 'x86_64'}--"
-            subprocess.run(
-                ["llc", "-filetype=obj", llc_triple, emi_ll_path, f"-o", emi_o_path]
-            )
-            for opt_level in experiment.opt_levels:
-                # opt_ll_path = os.path.join(tmpdir, f"{opt_level}.ll")
-                # asm_path = os.path.join(tmpdir, f"{opt_level}.s")
-                o_path = os.path.join(tmpdir, f"{opt_level}.o")
-                elf_path = os.path.join(tmpdir, f"{opt_level}.elf")
-                image_path = os.path.join(tmpdir, f"{opt_level}.bin")
-
-                # NOTE: for now, don't test middle end
-                # TODO: add opt testing support as cmdline switch
-                # subprocess.run(
-                #     [
-                #         "opt",
-                #         "-S",
-                #         f"--passes=default<O{opt_level}>",
-                #         "--disable-simplify-libcalls",
-                #         source_bc_path,
-                #         "-o",
-                #         opt_ll_path,
-                #     ]
-                # )
-
-                llc_args = [
-                    "llc",
-                    "-filetype=obj",
-                    f"-O{opt_level}",
-                    "--disable-simplify-libcalls",
-                    llc_triple,
-                    source_bc_path,  # opt_ll_path,
-                    "-o",
-                    o_path,
                 ]
-                if arch == "x86":
-                    llc_args += ["-mattr=+sse,+sse2", "--x86-asm-syntax=intel"]
-                subprocess.run(llc_args, stderr=subprocess.DEVNULL)
-                subprocess.run(
-                    ["ld", emi_o_path, o_path, "-o", elf_path],
-                    stderr=subprocess.DEVNULL,
-                )
+            )
 
-                try:
-                    fn_offset = get_sym_offset(elf_path, fn_name)
-                    emi_false_offset = get_sym_offset(elf_path, "emi_false", False)
-                except:
-                    rmdir(tmpdir)
-                    raise
+        subprocess.run(clang_args, stderr=subprocess.DEVNULL, check=True)
 
-                subprocess.run(
-                    ["objcopy", "-O", "binary", "-j", ".text", elf_path, image_path]
-                )
+    def _mutate(self, target_file: str, tmpdir: str, experiment: Experiment):
+        """Mutate bitcode file via IRFuzzer"""
+        assert target_file.endswith("out.bc")
+        irfuzzer_env = os.environ.copy()
+        irfuzzer_env["NUM_MUTATE"] = str(self.NUM_MUTATE)
+        subprocess.run(
+            ["MutatorDriver", target_file, str(experiment.randomizer.get())],
+            env=irfuzzer_env,
+            cwd=tmpdir,
+            stderr=subprocess.STDOUT if is_debug else subprocess.DEVNULL,
+            check=True,
+        )
 
-                with open(image_path, "rb") as image_file:
-                    image = image_file.read()
-                programs.append(
-                    Program(
-                        f"O{opt_level}",
-                        image,
-                        ret_ty,
-                        arg_tys,
-                        fn_offset,
-                        tmpdir,
-                        {emi_false_offset: b"\x00"},
-                    )
-                )
+    def _compile_elf(
+        self,
+        opt_level: str,
+        triple: str,
+        source: str,
+        output: str,
+        link_with: list[str] = [],
+    ):
+        stderr = subprocess.STDOUT if is_debug else subprocess.DEVNULL
 
-            return programs
+        source_type = ".bc" if ".bc" in source else ".ll"
+        optimized_source = f"{source.replace(source_type, '')}.{opt_level}{source_type}"
+        self._opt(opt_level, source, optimized_source)
+        obj_file = f"{source}.{opt_level}.o"
+        llc_args = [
+            "llc",
+            "-filetype=obj",
+            f"-O{opt_level}",
+            "--disable-simplify-libcalls",
+            f"-mtriple={triple}",
+            optimized_source,
+            "-o",
+            obj_file,
+        ]
+        if "x86" in triple:
+            llc_args += ["-mattr=+sse,+sse2", "--x86-asm-syntax=intel"]
+        subprocess.run(llc_args, stderr=stderr, check=True)
+        subprocess.run(
+            ["ld", *link_with, obj_file, "-o", output, "--warn-once"],
+            stderr=subprocess.DEVNULL,
+            check=True,
+        )
+
+    def _opt(self, opt_level: str, source: str, output: str):
+        subprocess.run(["opt", f"-O{opt_level}", source, "-o", output])
+
+    def _make_program(
+        self,
+        elf_path: str,
+        name: str,
+        fn_info: tuple[str, str, list[str]],
+        dir: str,
+        has_emi: bool,
+    ):
+        fn_name, ret_ty, arg_tys = fn_info
+        fn_offset = get_sym_offset(elf_path, fn_name)
+        if has_emi:
+            emi_false_offset = get_sym_offset(elf_path, "emi_false", False)
+        image_path = f"{elf_path}.bin"
+        subprocess.run(
+            ["objcopy", "-O", "binary", "-j", ".text", elf_path, image_path], check=True
+        )
+
+        with open(image_path, "rb") as image_file:
+            image = image_file.read()
+            consts = {emi_false_offset: b"\x00"} if has_emi else {}
+            return Program(
+                name,
+                image,
+                ret_ty,
+                arg_tys,
+                fn_offset,
+                dir,
+                consts,
+            )
 
     @property
     def name(self) -> str:
